move cfgspec/suite to config_schema and plug it in to config.py
